fixed test import, refactored home out of app.tsx
